Rearrange configurable parameters and update copy_page
Rearrange the configurable parameters, adding settings for the
preload cache line size and the preload offset. These can be
configured in the Makefile. Don't rely on the kernel's
L1_CACHE_BYTES definition.

Partly rewrite copy_page in order to preserve performance
when PRELOAD_CACHE_LINE_SIZE is equal to 32 on armv7.